Delete step by step link
Delete choose a supplier from a framework, step by step. How to use Find a framework to select a quality-checked supplier. Frameworks use pre-written terms and conditions that are approved by DfE. 
It's not possible to organise the content in this way. We also have no content or expertise to support this.